requirements.txt
fastapi
uvicorn[standard]
aiosqlite
python-dotenv
httpx
ijson
anthropic
pytest
pytest-asyncio
